Added TafListViewModel
