+++ HotelDemo/ViewModel/Components/TagCellViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  TagCellViewModel.swift
+//  HotelDemo
+//
+//  Created by Лаборатория on 08.09.2023.
+//
+
+import Foundation
+
+class TagCellViewModel: ObservableObject {
+
+    let title: String
+
+    init(title: String) {
+        self.title = title
+    }
+
+}
